DROTH-2954 Refactored LaneNumber
